Storage device & another bug fixes
There was a bug in the storage device transfer handler where
read/writing to a PhysicalRegionListBuffer was done incorrectly.

There was also a bug in the memory dumping code when printing the
hexadecimal values.

For some reason the console device was writing to stdout, when it should
always be writing to stderr. I'm unsure of when that was changed, but it
shouldn't have been.