Emulator: Configurable I/O interfaces + basic debugger
There are now configurable I/O interfaces for the console device and the
new basic debugger. These can be input/output on stdio (stdin/stdout),
output to a file, or input/output from a local TCP network port. By
default, the console device is on stdio, and the basic debugger is
disabled. The console device cannot be disabled.

The new basic debugger supports stepping instructions, basic breakpoints,
pause/continue execution (including interrupting with SIGINT), register
printing, memory region info dumping and reading from physical and virtual
memory. It is designed to be fully asynchronous, running on its own
independent thread. Execution is paused on startup when the debugger is
enabled.