Added a basic disassembler
It is very simple as most of the work is already done by libarch. This
simply just converts the data structures libarch outputs when given raw
machine code into human readible assembly code. It doesn't support raw
data or any directives as there is no way to detect these with flat
binaries. It will still output what it as disassembled so far if it
encounters an error.

Known bugs: The assembler seg faults when attempting to assemble a file
without any labels.