Instruction Cache + various other Emulator performance improvements
The biggest change in this commit is the introduction of an instruction
cache. The idea behind it is that <size> bytes are read from the MMU ahead
of time to optimise instruction times. This buffer is not aligned at
all. The current size is 256 bytes, but this might be tweaked in the
future.

The whole instruction decoding system had to be modified to
allow for this to happen. It was converted to use a stream-based buffer
instead of a standard dynamic buffer. The decoding system was also
modified to not re-read bytes for code simplicity anymore. The code
might be a big more complex and harder to follow, but it had a massive
performance boost. The Emulator performs 2x better with these changes.

Other performance-related changes or bug fixes were also made. These
include, but are not limited to, stack registers are now accessed
directly from the Stack data structure instead of constant syncing on
every instruction cycle, different types of register data structures
were made to allow for registers that need access or syncing checks
performed on them, the SimpleLinkedList got iterator support, allowing
for faster enumeration in the MMU, the test interrupt event was removed,
and all events now use the helper instead of manually adding to the
queue.

The Disassembler also had some bug fixes made to it with the
introduction of the changes. Various other smaller bug fixes or
improvements were also made.