Constant folding, added extra operand type and bug fixes
There is now basic constant folding in the assembler. Unary operations
like + or - are not supported, neither are operations including labels
or sub-labels. Expressions are evaluated left to right, with operator
precedence similar to languages like C. All expressions must be enclosed
in parentheses.

The double immediate version of the `[reg+offset]` format was re-added
so that operations like `[label + some number]` can still occur.

One major bug was that in the previous commit, I forgot to add support
for operands of type `MEMORY`, which is where just the base is present,
and it is an immediate.