Fixed mul/div, new ALU instructions, Emulator signal handling
`mul` and `div` have now been modified to operate using 3 operands,
which allows full size of results, and the `dividend` for division.

There are now 3 new instructions: `smul`, `sdiv`, and `xnor`. This adds
support for signed multiplication and division, and XNOR. The
multiplication and division instructions are almost the same as the
signed counterparts in terms of implementation. Some opcode reordering
also occurred because of these changes.

Some basic Emulator signal handling has been implemented so that
interrupts caused by instructions can be caught, e.g. SIGFPE because of
a bad division. It also handles a variety of other signals.