Proper instruction cache + other micro-optimisations
The main feature in this commit is that there is now a proper
instruction cache which caches the decoded information about
instructions. The old cache still exists, but it just caches memory used
for the instructions to save it from being read from the MMU. This
feature on average provides a 2-3 times speed up for loops.

A minor micro-optimisation was made that was forgotten about in the
previous commit where a fast path for the emulator itself checking
`SafeRegister`s.