Instruction cache bug fixes + debug print fixes
There were 2 main instruction cache bugs that are fixed in this, one
related to an overflow of the cache, and the other related to LibArch
not having a big enough global buffer for the instruction data when the
cache is used.

The formatting of debug printing of operands was also fixed, along with
a tiny bit of cleanup in the CMakeLists.txt for the Emulator.