Added a basic Exec library
The new LibExec is intended for interacting with executable file formats
other than basic binary files. Currently it only supports writing some
aspects of ELF files including program headers and sections of type
PROGBITS.

There is also a ELFTest example to test the new library and show how it
works. This is only temporary until this library is used in main parts
of the project.